Simpel matching

Flexible and fast matching functions.

Usage

cc_match(
  cases,
  controls,
  by,
  id,
  no_controls,
  replace = TRUE,
  seed = NULL,
  return_case_values = FALSE,
  verbose = TRUE
)

Arguments

cases

A data.frame including cases.

controls

A data.frame including potential controls.

by

A list of matching factors and their accompanying matching criteria either supplied as a function or as 'exact', if exact matching is requested for a specific variable. The matching function should take two arguments as input where the first argument is the value of the case and the second argument is the value of the control. The function function(x,y) abs(x - y) < 1 can for example be used to match cases to compactors with a maximal age difference of one year. The function function(x,y) y >= x can be used for concurrent matching, i.e., requiering the control to be alive at the time of matching.

Example:

list(sex  = 'exact',
     age  = function(x,y) abs(x - y) < 1,
     time = function(x,y) y >= x)

id

Name of the variable that identifies unique observations.

no_controls

The number of controls that should be matched to each case.

replace

Logical indicator if sampling of controls should be done with or without replacement. Default: FALSE.

seed

Optional seed used for the matching. Useful for reproducibility.

return_case_values

If TRUE the case value of the matching variable will be returned in the output data set, e.g., if you match on time, the controls in the output dataset will have a variable case_time which has the value of the case's time variable. This can be useful if you would like to start follow-up at the time of matching in your later analysis.

verbose

Logical indicator if default checks should be printed. Default: TRUE.

Value

A dataset of matched cases and controls.

id

Individuals unique id

case

Case indicator

riskset

Riskset number/identifier

Details

The function returns a warning if the number of available controls is smaller than the number of controls requested for each case. In this case all available controls are matched to the case.

If there are no matched available for a particular case, the case will be removed from the output dataset and a warning will be shown indicating the id of the case that has been removed.

If you want to match exact on a variable, please use the 'exact' indicator as shown above. This will improve the computation time.

Examples

require(rstpm2)
#> Loading required package: rstpm2
#> Loading required package: survival
#> Loading required package: splines
#> 
#> Attaching package: ‘rstpm2’
#> The following object is masked from ‘package:survival’:
#> 
#>     colon
cc_match(cases    = brcancer[brcancer$hormon == 1, ],
         controls = brcancer[brcancer$hormon == 0, ],
         id = "id",
         by = list(x2 = "exact",
                   rectime = function(x, y) y >= x),
         no_controls = 1,
         replace = TRUE)
#> Warning: ✖ No controls available for 236.
#> ℹ I'll remove this control form the output dataset.
#> Warning: ✖ No controls available for 272.
#> ℹ I'll remove this control form the output dataset.
#> Warning: ✖ No controls available for 275.
#> ℹ I'll remove this control form the output dataset.
#> Warning: ✖ No controls available for 468.
#> ℹ I'll remove this control form the output dataset.
#> ----------------------------------------
#> No. of comparators: 242 
#> No. of unique comparators: 124 
#> Proportion of unique compartors: 51.2 %
#> ----------------------------------------
#> 
#>      id case riskset
#> 1     2    1       1
#> 2   432    0       1
#> 3     3    1       2
#> 4   240    0       2
#> 5     4    1       3
#> 6    46    0       3
#> 7     7    1       4
#> 8   279    0       4
#> 9    11    1       5
#> 10  472    0       5
#> 11   12    1       6
#> 12  121    0       6
#> 13   13    1       7
#> 14   46    0       7
#> 15   15    1       8
#> 16   92    0       8
#> 17   18    1       9
#> 18  163    0       9
#> 19   19    1      10
#> 20   46    0      10
#> 21   24    1      11
#> 22   22    0      11
#> 23   27    1      12
#> 24   76    0      12
#> 25   29    1      13
#> 26  109    0      13
#> 27   33    1      14
#> 28    8    0      14
#> 29   36    1      15
#> 30  584    0      15
#> 31   44    1      16
#> 32  294    0      16
#> 33   45    1      17
#> 34  174    0      17
#> 35   48    1      18
#> 36  208    0      18
#> 37   49    1      19
#> 38  266    0      19
#> 39   51    1      20
#> 40   21    0      20
#> 41   54    1      21
#> 42  655    0      21
#> 43   57    1      22
#> 44  573    0      22
#> 45   58    1      23
#> 46   10    0      23
#> 47   60    1      24
#> 48   22    0      24
#> 49   61    1      25
#> 50   10    0      25
#> 51   65    1      26
#> 52  472    0      26
#> 53   73    1      27
#> 54  174    0      27
#> 55   85    1      28
#> 56  518    0      28
#> 57   86    1      29
#> 58  522    0      29
#> 59   91    1      30
#> 60  126    0      30
#> 61   93    1      31
#> 62  182    0      31
#> 63   99    1      32
#> 64  269    0      32
#> 65  115    1      33
#> 66  229    0      33
#> 67  118    1      34
#> 68  301    0      34
#> 69  125    1      35
#> 70  163    0      35
#> 71  127    1      36
#> 72  531    0      36
#> 73  129    1      37
#> 74  192    0      37
#> 75  130    1      38
#> 76  522    0      38
#> 77  131    1      39
#> 78   46    0      39
#> 79  135    1      40
#> 80  100    0      40
#> 81  136    1      41
#> 82  605    0      41
#> 83  137    1      42
#> 84  363    0      42
#> 85  144    1      43
#> 86  337    0      43
#> 87  145    1      44
#> 88  403    0      44
#> 89  146    1      45
#> 90  451    0      45
#> 91  148    1      46
#> 92  198    0      46
#> 93  153    1      47
#> 94  246    0      47
#> 95  155    1      48
#> 96  112    0      48
#> 97  157    1      49
#> 98   21    0      49
#> 99  177    1      50
#> 100   8    0      50
#> 101 178    1      51
#> 102 283    0      51
#> 103 180    1      52
#> 104 376    0      52
#> 105 181    1      53
#> 106 377    0      53
#> 107 183    1      54
#> 108 377    0      54
#> 109 184    1      55
#> 110 451    0      55
#> 111 185    1      56
#> 112 433    0      56
#> 113 186    1      57
#> 114  46    0      57
#> 115 188    1      58
#> 116  46    0      58
#> 117 191    1      59
#> 118 451    0      59
#> 119 194    1      60
#> 120  10    0      60
#> 121 196    1      61
#> 122 266    0      61
#> 123 200    1      62
#> 124 407    0      62
#> 125 203    1      63
#> 126 279    0      63
#> 127 213    1      64
#> 128 266    0      64
#> 129 215    1      65
#> 130 438    0      65
#> 131 218    1      66
#> 132   5    0      66
#> 133 220    1      67
#> 134 240    0      67
#> 135 221    1      68
#> 136 143    0      68
#> 137 223    1      69
#> 138 182    0      69
#> 139 226    1      70
#> 140 474    0      70
#> 141 227    1      71
#> 142  37    0      71
#> 143 228    1      72
#> 144 621    0      72
#> 145 233    1      73
#> 146 150    0      73
#> 147 234    1      74
#> 148 168    0      74
#> 149 239    1      76
#> 150 584    0      76
#> 151 241    1      77
#> 152 377    0      77
#> 153 242    1      78
#> 154 474    0      78
#> 155 243    1      79
#> 156  46    0      79
#> 157 245    1      80
#> 158  16    0      80
#> 159 248    1      81
#> 160 425    0      81
#> 161 250    1      82
#> 162 192    0      82
#> 163 251    1      83
#> 164 279    0      83
#> 165 254    1      84
#> 166 522    0      84
#> 167 257    1      85
#> 168 412    0      85
#> 169 258    1      86
#> 170 549    0      86
#> 171 259    1      87
#> 172 540    0      87
#> 173 261    1      88
#> 174 377    0      88
#> 175 262    1      89
#> 176  87    0      89
#> 177 264    1      90
#> 178  47    0      90
#> 179 273    1      92
#> 180 279    0      92
#> 181 276    1      94
#> 182  39    0      94
#> 183 277    1      95
#> 184 402    0      95
#> 185 278    1      96
#> 186 182    0      96
#> 187 281    1      97
#> 188  94    0      97
#> 189 282    1      98
#> 190 456    0      98
#> 191 284    1      99
#> 192 540    0      99
#> 193 286    1     100
#> 194 352    0     100
#> 195 287    1     101
#> 196 266    0     101
#> 197 288    1     102
#> 198 294    0     102
#> 199 291    1     103
#> 200 659    0     103
#> 201 293    1     104
#> 202 377    0     104
#> 203 295    1     105
#> 204 518    0     105
#> 205 302    1     106
#> 206 432    0     106
#> 207 304    1     107
#> 208 605    0     107
#> 209 308    1     108
#> 210 150    0     108
#> 211 314    1     109
#> 212 240    0     109
#> 213 316    1     110
#> 214 474    0     110
#> 215 318    1     111
#> 216  10    0     111
#> 217 319    1     112
#> 218 576    0     112
#> 219 323    1     113
#> 220 347    0     113
#> 221 329    1     114
#> 222 325    0     114
#> 223 330    1     115
#> 224 561    0     115
#> 225 332    1     116
#> 226 522    0     116
#> 227 333    1     117
#> 228 216    0     117
#> 229 336    1     118
#> 230 397    0     118
#> 231 338    1     119
#> 232 341    0     119
#> 233 340    1     120
#> 234  37    0     120
#> 235 344    1     121
#> 236 290    0     121
#> 237 348    1     122
#> 238 133    0     122
#> 239 354    1     123
#> 240 407    0     123
#> 241 357    1     124
#> 242   8    0     124
#> 243 358    1     125
#> 244 182    0     125
#> 245 362    1     126
#> 246 371    0     126
#> 247 365    1     127
#> 248  53    0     127
#> 249 368    1     128
#> 250 371    0     128
#> 251 369    1     129
#> 252  47    0     129
#> 253 370    1     130
#> 254 377    0     130
#> 255 385    1     131
#> 256 456    0     131
#> 257 387    1     132
#> 258 672    0     132
#> 259 388    1     133
#> 260  52    0     133
#> 261 390    1     134
#> 262 160    0     134
#> 263 391    1     135
#> 264 271    0     135
#> 265 398    1     136
#> 266 381    0     136
#> 267 404    1     137
#> 268 632    0     137
#> 269 405    1     138
#> 270 522    0     138
#> 271 408    1     139
#> 272 299    0     139
#> 273 413    1     140
#> 274 605    0     140
#> 275 414    1     141
#> 276 386    0     141
#> 277 415    1     142
#> 278 285    0     142
#> 279 416    1     143
#> 280 271    0     143
#> 281 418    1     144
#> 282 573    0     144
#> 283 419    1     145
#> 284 527    0     145
#> 285 421    1     146
#> 286 655    0     146
#> 287 429    1     147
#> 288 522    0     147
#> 289 430    1     148
#> 290 650    0     148
#> 291 431    1     149
#> 292 409    0     149
#> 293 436    1     150
#> 294  14    0     150
#> 295 437    1     151
#> 296 328    0     151
#> 297 440    1     152
#> 298 174    0     152
#> 299 446    1     153
#> 300 584    0     153
#> 301 450    1     154
#> 302 605    0     154
#> 303 453    1     155
#> 304  52    0     155
#> 305 454    1     156
#> 306 650    0     156
#> 307 455    1     157
#> 308 112    0     157
#> 309 459    1     158
#> 310 684    0     158
#> 311 460    1     159
#> 312 377    0     159
#> 313 461    1     160
#> 314  84    0     160
#> 315 462    1     161
#> 316 375    0     161
#> 317 466    1     162
#> 318  23    0     162
#> 319 467    1     163
#> 320 306    0     163
#> 321 469    1     165
#> 322 631    0     165
#> 323 470    1     166
#> 324  46    0     166
#> 325 475    1     167
#> 326 201    0     167
#> 327 476    1     168
#> 328 376    0     168
#> 329 477    1     169
#> 330 371    0     169
#> 331 478    1     170
#> 332 263    0     170
#> 333 479    1     171
#> 334 572    0     171
#> 335 481    1     172
#> 336 402    0     172
#> 337 483    1     173
#> 338 474    0     173
#> 339 484    1     174
#> 340 163    0     174
#> 341 487    1     175
#> 342  62    0     175
#> 343 489    1     176
#> 344 660    0     176
#> 345 491    1     177
#> 346 334    0     177
#> 347 492    1     178
#> 348 187    0     178
#> 349 493    1     179
#> 350 549    0     179
#> 351 495    1     180
#> 352  94    0     180
#> 353 496    1     181
#> 354 402    0     181
#> 355 500    1     182
#> 356 520    0     182
#> 357 502    1     183
#> 358  25    0     183
#> 359 503    1     184
#> 360  96    0     184
#> 361 504    1     185
#> 362 163    0     185
#> 363 508    1     186
#> 364 292    0     186
#> 365 510    1     187
#> 366 672    0     187
#> 367 514    1     188
#> 368 193    0     188
#> 369 523    1     189
#> 370 290    0     189
#> 371 524    1     190
#> 372 474    0     190
#> 373 526    1     191
#> 374 280    0     191
#> 375 529    1     192
#> 376 280    0     192
#> 377 532    1     193
#> 378 174    0     193
#> 379 536    1     194
#> 380 516    0     194
#> 381 538    1     195
#> 382   8    0     195
#> 383 539    1     196
#> 384 520    0     196
#> 385 544    1     197
#> 386 174    0     197
#> 387 548    1     198
#> 388 133    0     198
#> 389 550    1     199
#> 390  89    0     199
#> 391 552    1     200
#> 392  39    0     200
#> 393 553    1     201
#> 394 309    0     201
#> 395 560    1     202
#> 396 474    0     202
#> 397 562    1     203
#> 398 432    0     203
#> 399 564    1     204
#> 400  96    0     204
#> 401 568    1     205
#> 402 660    0     205
#> 403 571    1     206
#> 404 434    0     206
#> 405 574    1     207
#> 406 175    0     207
#> 407 581    1     208
#> 408  34    0     208
#> 409 582    1     209
#> 410 111    0     209
#> 411 585    1     210
#> 412 577    0     210
#> 413 588    1     211
#> 414 229    0     211
#> 415 590    1     212
#> 416 584    0     212
#> 417 597    1     213
#> 418 386    0     213
#> 419 598    1     214
#> 420 116    0     214
#> 421 599    1     215
#> 422 133    0     215
#> 423 603    1     216
#> 424 150    0     216
#> 425 608    1     217
#> 426 143    0     217
#> 427 611    1     218
#> 428 107    0     218
#> 429 614    1     219
#> 430 410    0     219
#> 431 615    1     220
#> 432  72    0     220
#> 433 617    1     221
#> 434 133    0     221
#> 435 620    1     222
#> 436 540    0     222
#> 437 627    1     223
#> 438 565    0     223
#> 439 629    1     224
#> 440 240    0     224
#> 441 630    1     225
#> 442 456    0     225
#> 443 635    1     226
#> 444  14    0     226
#> 445 636    1     227
#> 446 168    0     227
#> 447 641    1     228
#> 448 451    0     228
#> 449 642    1     229
#> 450 266    0     229
#> 451 643    1     230
#> 452 520    0     230
#> 453 644    1     231
#> 454 631    0     231
#> 455 647    1     232
#> 456 112    0     232
#> 457 649    1     233
#> 458 363    0     233
#> 459 653    1     234
#> 460 403    0     234
#> 461 654    1     235
#> 462 375    0     235
#> 463 656    1     236
#> 464 573    0     236
#> 465 657    1     237
#> 466  94    0     237
#> 467 662    1     238
#> 468 271    0     238
#> 469 665    1     239
#> 470 664    0     239
#> 471 670    1     240
#> 472 280    0     240
#> 473 671    1     241
#> 474 280    0     241
#> 475 675    1     242
#> 476 159    0     242
#> 477 677    1     243
#> 478 386    0     243
#> 479 679    1     244
#> 480 474    0     244
#> 481 681    1     245
#> 482 187    0     245
#> 483 683    1     246
#> 484  84    0     246