Simpel matching

Flexible and fast matching functions.

Usage

matching(
  data,
  case_indicator,
  id,
  matching_factors,
  no_controls,
  replace = FALSE,
  seed = NULL,
  verbose = TRUE
)

Arguments

data

A data.frame including both cases and potential controls.

case_indicator

Name of the variable that indicates if an observation is a case or a controls, codes as 1 if case and 0 if control.

id

Name of the variable that identifies unique observations.

matching_factors

A list of matching factors and their accompanying matching criteria either supplied as a function or as 'exact', if exact matching is requested for a specific variable. The matching function should take two arguments as input where the first argument is the value of the case and the second argument is the value of the control. The function function(x,y) abs(x - y) < 1 can for example be used to match cases to compactors with a maximal age difference of one year. The function function(x,y) y >= x can be used for concurrent matching, i.e., requiering the control to be alive at the time of matching.

Example:

list(sex  = 'exact',
     age  = function(x,y) abs(x - y) < 1,
     time = function(x,y) y >= x)

no_controls

The number of controls that should be matched to each case.

replace

Logical indicator if sampling of controls should be done with or without replacement. Default: FALSE.

seed

Optional seed used for the matching. Usefull for reproducibility.

verbose

Logical indicator if default checks should be printed. Default: TRUE.

Value

A dataset of matched cases and controls.

id

Individuals unique id

case

Case indicator

riskset

Riskset number/identifier

All the variables included in the supplied dataset are also returned.

Details

The function returns a warning if the number of available controls is smaller than the number of controls requested for each case. In this case all availabe controls are matched to the case.

If there are no matched availabe for a particular case, the case will be removed from the output dataset and a warning will be shown indicating the id of the case that has been removed.

If you want to match exact on a variable, please use the 'exact' indicator as shown above. This will improve the computation time.

Examples

require(rstpm2)

matching(brcancer,
         case_indicator = "hormon",
         id = "id",
         matching_factors = list(x2 = "exact",
                                 rectime = function(x, y) y >= x),
         no_controls = 1)
#> Warning: ✖ No controls available for 236.
#> ℹ I'll remove this control form the output dataset.
#> Warning: ✖ No controls available for 272.
#> ℹ I'll remove this control form the output dataset.
#> Warning: ✖ No controls available for 275.
#> ℹ I'll remove this control form the output dataset.
#> Warning: ✖ No controls available for 468.
#> ℹ I'll remove this control form the output dataset.
#> Error: Object 'hormon' not found amongst [id, case, riskset, x1, x3, x4, x5, x6, x7, censrec, ...]